[🚚RENAME] 폴더명 및 경로 수정
